+++ crates/starknet-devnet-server/src/rpc_handler.rs
@@ -5,7 +5,7 @@ use axum::extract::State;
 use axum::Json;
 use futures::{future, FutureExt};
 use serde::de::DeserializeOwned;
-use tracing::{error, trace, warn};
+use tracing::{trace, warn};
 
 use crate::rpc_core::error::RpcError;
 use crate::rpc_core::request::{Request, RpcCall, RpcMethodCall};
@@ -32,36 +32,7 @@ pub trait RpcHandler: Clone + Send + Sync + 'static {
     ///
     /// **Note**: override this function if the expected `Request` deviates from `{ "method" :
     /// "<name>", "params": "<params>" }`
-    async fn on_call(&mut self, call: RpcMethodCall) -> RpcResponse {
-        trace!(target: "rpc",  id = ?call.id , method = ?call.method, "received method call");
-        let RpcMethodCall { method, params, id, .. } = call.clone();
-
-        let params: serde_json::Value = params.into();
-        let deserializable_call = serde_json::json!({
-            "method": &method,
-            "params": params
-        });
-
-        match serde_json::from_value::<Self::Request>(deserializable_call) {
-            Ok(req) => {
-                let result = self.on_request(req, call).await;
-                RpcResponse::new(id, result)
-            }
-            Err(err) => {
-                let err = err.to_string();
-                // since JSON-RPC specification requires returning a Method Not Found error,
-                // we apply a hacky way to induce this - checking the stringified error message
-                let distinctive_error = format!("unknown variant `{method}`");
-                if err.contains(&distinctive_error) {
-                    error!(target: "rpc", ?method, "failed to deserialize method due to unknown variant");
-                    RpcResponse::new(id, RpcError::method_not_found())
-                } else {
-                    error!(target: "rpc", ?method, ?err, "failed to deserialize method");
-                    RpcResponse::new(id, RpcError::invalid_params(err))
-                }
-            }
-        }
-    }
+    async fn on_call(&mut self, call: RpcMethodCall) -> RpcResponse;
 }
 
 /// Handles incoming JSON-RPC Request
