+++ crates/starknet-devnet/tests/general_integration_tests.rs
@@ -87,10 +87,11 @@ mod general_integration_tests {
                 // expected port added after spawning; determined by port-acquiring logic
                 "timeout": 121,
                 "request_body_size_limit": 1000,
+                "restricted_methods": null,
             },
             "block_generation_on": "demand",
             "lite_mode": false,
-            "disable_account_impersonation": false,
+            "disable_account_impersonation": false
         });
 
         let devnet = BackgroundDevnet::spawn_with_additional_args(&[
